refactor: extract shared profile-bee-symbols crate
Create profile-bee-symbols as a shared library crate used by both
profile-bee and symbol-server:

- fileid.rs: htlhash FileId computation (SHA-256 of head+tail+length),
  compatible with devfiler/Elastic format
- extract.rs: ELF symbol extraction from .symtab/.dynsym with
  Rust/C++ demangling, DWARF enrichment stub for future work

Deduplicates htlhash computation that was previously inlined in
profile-bee's otlp.rs and duplicated in symbol-server. Both crates
now depend on profile-bee-symbols for consistent behavior.

This enables future optimizations where profile-bee can optionally
pre-extract symbols (since it already runs blazesym) and POST them
in symbfile format directly, reducing symbol-server's workload.
