+++ profile-bee/Cargo.toml
@@ -21,6 +21,7 @@ include = [
 aya = { package = "profile-bee-aya", version = "0.13.2" }
 aya-log = { package = "profile-bee-aya-log", version = "0.2.2" }
 profile-bee-common = { version = "0.3.12", path = "../profile-bee-common", features=["user"] }
+profile-bee-symbols = { path = "../profile-bee-symbols" }
 profile-bee-tui = { version = "0.3.12", path = "../profile-bee-tui", optional = true }
 anyhow = "1.0.42"
 clap = { version = "4.0", features = ["derive"] }
@@ -49,7 +50,6 @@ flate2 = "1"
 strip-ansi-escapes = "0.2"
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"] }
 parking_lot = "0.12"
-sha2 = "0.10"
 aws-sdk-codeguruprofiler = { version = "1", optional = true }
 aws-config = { version = "1", features = ["behavior-version-latest"], optional = true }
 tonic = { version = "0.12", features = ["gzip", "transport"], optional = true }
